requirements.txt
discord.py[voice]>=2.0.0
yt-dlp>=2023.3.4
python-dotenv>=0.19.0
async-timeout>=4.0.0
pynacl>=1.5.0
ffmpeg-python
asyncio
flask